Moved WebSocket files to `networking` folder.
